implemented the ImageStore

+++ src/main/java/me/blvckbytes/bluemap_markers/model/MapImage.java
@@ -3,5 +3,5 @@ package me.blvckbytes.bluemap_markers.model;
 public record MapImage(
   String fileName,
   String absolutePath,
-  int fileSizeInBytes
+  long fileSizeInBytes
 ) {}
